+++ src/Frontend/Dfe.Complete/Constants/RouteConstants.cs
@@ -13,6 +13,7 @@
         // Your team projects
         public const string TeamProjectsInProgress = "/projects/team/in-progress";
         public const string TeamProjectsNew = "/projects/team/new";
+        public const string TeamProjectsUsers = "/projects/team/users";
         public const string TeamProjectsHandedOver = "/projects/team/handed-over";
         public const string TeamProjectsCompleted = "/projects/team/completed";
         public const string TeamProjectsUnassigned = "/projects/team/unassigned";